+++ src/view/film-card.js
@@ -1,10 +1,10 @@
 export const createFilmCardTemplate = (film) => {
 /*  const {description} = film;*/
-  let {title, genre, poster, description} = film;
+  let {title, rating, genre, poster, description} = film;
 
   return `<article class="film-card">
           <h3 class="film-card__title">${title}</h3>
-          <p class="film-card__rating">8.3</p>
+          <p class="film-card__rating">${rating.toFixed(1)}</p>
           <p class="film-card__info">
             <span class="film-card__year">1929</span>
             <span class="film-card__duration">1h 55m</span>
